apps/api_fastapi/pyproject.toml
[project]
name = "api-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "fastapi>=0.117.1",
    "langchain-core>=0.3.76",
    "langchain-google-genai>=2.1.12",
    "langgraph>=0.6.7",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "uvicorn>=0.37.0",
]
